pyproject.toml
[tool.poetry]
name = "qualidade-em-projeto-de-dados"
version = "0.1.0"
description = ""
authors = ["Calel <borgescalel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mkdocs = "^1.6.1"
mkdocs-mermaid2-plugin = "^1.2.1"
mkdocs-material = "^9.6.14"
mkdocstrings = {extras = ["python"], version = "^0.29.1"}
taskipy = "^1.14.1"
isort = "^6.0.1"
black = "^25.1.0"
pytest = "^8.3.5"
pydantic = "^2.11.5"
pandera = "^0.24.0"
sqlalchemy = "^2.0.41"
python-dotenv = "^1.1.0"
psycopg2 = "^2.9.10"
pandas = "^2.2.3"
frictionless = "^5.18.1"
mkdocstrings-python = "^1.16.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
format = """
isort .
black .
"""
kill = "kill -9 $(lsof -t -i :8000)"
test = "pytest -v"
run = """
python3 app/main.py
"""

doc = "mkdocs serve"